:construction: Expose inference again
